refactor : 프로젝트 구조 재정의 및 클래스명 변경하여 유지보수성 및 가독성 향상
