+++ src/main/java/racingGame/Business/Process.java
@@ -1,0 +1,5 @@
+package racingGame.Business;
+
+public class Process {
+
+}
