+++ src/main/java/racingGame/GameApplication.java
@@ -1,0 +1,8 @@
+package racingGame;
+
+public class GameApplication {
+    
+    public static void main(String[] args) {
+    
+    }
+}
