--- src/main/java/racingGame/Gameprocess.java
@@ -1,5 +1,0 @@
-package racingGame;
-
-public class Gameprocess {
-
-}
