--- src/test/java/racingGame/GameProcessTest.java
@@ -1,75 +1,0 @@
-package racingGame;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import racingGame.model.Car;
-
-class GameProcessTest {
-    
-    Gameprocess gp;
-    
-    @BeforeEach
-    void setUp() {
-        gp = new Gameprocess();
-    }
-    
-    @Test
-    void carGenerateResultNullTest() {
-        gp.run(-1,1);
-        assertThatThrownBy(() -> {
-            gp.generateCar(-1);
-        }).isInstanceOf(IllegalArgumentException.class)
-            .hasMessage("참가시킬 자동차 수가 없습니다.");
-    }
-
-    @Test
-    void gameTurnSetNullTest() {
-        gp.run(1,-1);
-        assertThatThrownBy(() -> {
-            gp.generateCar(-1);
-        }).isInstanceOf(IllegalArgumentException.class)
-            .hasMessage("몇번의 게임을 할지 값을 입력해주세요 ");
-    }
-
-    @Test
-    void carGenerateResultTest() {
-        gp.run(3,1);
-        assertThat(gp.joinCars().size).isEqual(3);
-    }
-
-    @Test
-    void gameNextTurnTest() {
-        gp.run(1,2);
-        assertThat(gp.nowTurn()).isEqual(2);
-    }
-
-    @Test
-    void carForwardTest() {
-        gp.run(1, 10);
-        while(true) {
-            gp.nextTurn();
-            if(gp.joinCars.get(0).loaction > 0) {
-                break;
-            }
-        }
-
-        assertThat(gp.joinCars.get(0).loaction).isNotEqual(0);
-    }
-
-    @Test
-    void carsForwardTest() {
-        gp.run(3, 10);
-        while(true) {
-            gp.nextTurn();
-            if(gp.joinCars.get(0).loaction > 2) {
-                break;
-            }
-        }
-        for(Car car : gp.joinCars){
-            assertThat(car.loaction).isNotEqual(0);
-        }
-    }
-}
